git-svn-id: http://svn.code.sf.net/p/genoviz/code/branches/das2_manager@4263 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ genoviz_sdk/demo/src/genoviz/demo/adapter/AnnotationAdapter.java
@@ -70,10 +70,10 @@ public class AnnotationAdapter implements NeoDataAdapterI {
 
 		GlyphI gl;
 		if (start <= end) {
-			gl = factory.makeGlyph((float)start, (float)(end+1));
+			gl = factory.makeGlyph(start, end+1.0f);
 		}
 		else {
-			gl = factory.makeGlyph((float)(start+1), (float)end);
+			gl = factory.makeGlyph(start+1.0f, end);
 		}
 		return gl;
 	}
