+++ genoviz_sdk/demo/src/genoviz/demo/adapter/ScoredAnnotAdapter.java
@@ -81,10 +81,10 @@ public class ScoredAnnotAdapter implements NeoDataAdapterI {
 		factory.setColor(col);
 		GlyphI gl;
 		if (start <= end) {
-			gl = factory.makeGlyph((float)start, (float)(end+1));
+			gl = factory.makeGlyph(start, end+1.0f);
 		}
 		else {
-			gl = factory.makeGlyph((float)(start+1), (float)end);
+			gl = factory.makeGlyph(start+1.0f, end);
 		}
 		return gl;
 	}
