+++ genoviz_sdk/demo/src/genoviz/demo/adapter/SequenceAnnotAdapter.java
@@ -58,10 +58,10 @@ public class SequenceAnnotAdapter implements NeoDataAdapterI {
 
 		GlyphI gl;
 		if (start <= end) {
-			gl = factory.makeGlyph((float)start, (float)(end+1));
+			gl = factory.makeGlyph(start, end+1.0f);
 		}
 		else {
-			gl = factory.makeGlyph((float)(start+1), (float)end);
+			gl = factory.makeGlyph(start+1.0f, end);
 		}
 		((SequenceGlyph)gl).setResidues(residues);
 		return gl;
